+++ src/Libraries/RevitNodes/Elements/Viewport.cs
@@ -45,9 +45,20 @@ namespace Revit.Elements
             SafeInit(() => InitViewport(viewport));
         }
 
+        /// <summary>
+        /// Create a new Viewport by Sheet, View and Location
+        /// </summary>
+        /// <param name="sheet"></param>
+        /// <param name="view"></param>
+        /// <param name="location"></param>
+        private Viewport(Sheet sheet, Revit.Elements.Views.View view, Autodesk.DesignScript.Geometry.Point location)
+        {
+            SafeInit(() => InitViewport(sheet, view, location));
+        }
+
         #endregion
 
-        #region Private constructors
+        #region Helpers for Private constructors
 
         /// <summary>
         /// Initialize a Viewport element
@@ -56,6 +67,47 @@ namespace Revit.Elements
         private void InitViewport(Autodesk.Revit.DB.Viewport viewport)
         {
             InternalSetViewport(viewport);
+        }
+
+        /// <summary>
+        /// Initialize a Viewport element
+        /// </summary>
+        /// <param name="sheet"></param>
+        /// <param name="view"></param>
+        /// <param name="location"></param>
+        private void InitViewport(Sheet sheet, Revit.Elements.Views.View view, Autodesk.DesignScript.Geometry.Point location)
+        {
+            ElementId sheetId = sheet.InternalView.Id;
+            ElementId viewId = view.InternalView.Id;
+            XYZ viewLocation = GeometryPrimitiveConverter.ToRevitType(location);
+
+            TransactionManager.Instance.EnsureInTransaction(Document);
+
+            var viewportElement = ElementBinder.GetElementFromTrace<Autodesk.Revit.DB.Viewport>(Document);
+
+            if (viewportElement == null)
+            {
+                if (!Autodesk.Revit.DB.Viewport.CanAddViewToSheet(Document, sheetId, viewId))
+                    throw new InvalidOperationException(Properties.Resources.ViewAlreadyPlacedOnSheet);
+                viewportElement = Autodesk.Revit.DB.Viewport.Create(Document, sheetId, viewId, viewLocation);
+            }
+            else
+            {
+                if (!Autodesk.Revit.DB.Viewport.CanAddViewToSheet(Document, sheetId, viewId))
+                {
+                    viewportElement.SetBoxCenter(viewLocation);
+                }                    
+                else
+                {
+                    viewportElement = Autodesk.Revit.DB.Viewport.Create(Document, sheetId, viewId, viewLocation);
+                }
+            }
+
+            InternalSetViewport(viewportElement);
+
+            TransactionManager.Instance.TransactionTaskDone();
+
+            ElementBinder.CleanupAndSetElementForTrace(Document, this.InternalElement);
         }
 
         #endregion
@@ -76,7 +128,76 @@ namespace Revit.Elements
         #endregion
 
         #region Public properties
+        /// <summary>
+        /// Get ViewSheet on which the viewport appears
+        /// </summary>
+        public Revit.Elements.Views.Sheet Sheet
+        {
+            get
+            {
+                var sheet = (Autodesk.Revit.DB.ViewSheet)Document.GetElement(InternalViewport.SheetId);
+                return (Revit.Elements.Views.Sheet)ElementWrapper.ToDSType(sheet, true);
+            }
+        }
 
+        /// <summary>
+        /// Get the associated View 
+        /// </summary>
+        public Revit.Elements.Views.View View
+        {
+            get
+            {
+                var view = (Autodesk.Revit.DB.View)Document.GetElement(InternalViewport.ViewId);
+                return (Revit.Elements.Views.View) ElementWrapper.ToDSType(view, true);
+            }
+        }
+
+        /// <summary>
+        /// Returns the center of the outline of the viewport on the sheet, excluding the viewport label. 
+        /// </summary>
+        public Autodesk.DesignScript.Geometry.Point BoxCenter
+        {
+            get
+            {
+                return InternalViewport.GetBoxCenter().ToPoint();
+            }
+        }
+
+        /// <summary>
+        /// Moves this viewport so that the center of the box outline (excluding the viewport label) is at a given point.
+        /// </summary>
+        /// <param name="point"></param>
+        public Viewport SetBoxCenter(Autodesk.DesignScript.Geometry.Point point)
+        {
+            TransactionManager.Instance.EnsureInTransaction(Application.Document.Current.InternalDocument);
+            this.InternalViewport.SetBoxCenter(point.ToXyz());
+            TransactionManager.Instance.TransactionTaskDone();
+            return this;
+        }
+
+        /// <summary>
+        /// Returns the outline of the viewport on the sheet, excluding the viewport label.
+        /// </summary>
+        public BoundingBox BoxOutline
+        {
+            get
+            {
+                var outline = InternalViewport.GetBoxOutline();
+                return BoundingBox.ByCorners(outline.MinimumPoint.ToPoint(), outline.MaximumPoint.ToPoint());
+            }
+        }
+
+        /// <summary>
+        /// Gets the outline viewport's label on the sheet.
+        /// </summary>
+        public BoundingBox LabelOutline
+        {
+            get
+            {
+                var outline = InternalViewport.GetLabelOutline();
+                return BoundingBox.ByCorners(outline.MinimumPoint.ToPoint(), outline.MaximumPoint.ToPoint());
+            }
+        }
         #endregion
 
         #region Public static constructors
@@ -90,19 +211,10 @@ namespace Revit.Elements
         /// <returns>The new Viewport.</returns>
         public static Viewport BySheetViewLocation(Sheet sheet, Revit.Elements.Views.View view, Autodesk.DesignScript.Geometry.Point location)
         {
-            ElementId sheetId = sheet.InternalView.Id;
-            ElementId viewId = view.InternalView.Id;
-            XYZ viewLocation = GeometryPrimitiveConverter.ToRevitType(location);
-
-            if (!Autodesk.Revit.DB.Viewport.CanAddViewToSheet(Document, sheetId, viewId))
-                throw new InvalidOperationException(Properties.Resources.ViewAlreadyPlacedOnSheet);
             if (IsViewEmpty(view.InternalView))
                 throw new InvalidOperationException(Properties.Resources.EmptyView);
 
-            TransactionManager.Instance.EnsureInTransaction(Document);
-            var viewport = Autodesk.Revit.DB.Viewport.Create(Document, sheetId, viewId, viewLocation);
-            TransactionManager.Instance.TransactionTaskDone();
-            return viewport.ToDSType(true) as Viewport;
+            return new Viewport(sheet, view, location);
         }
 
         #endregion